\name{tumor-package}
\alias{tumor-package}
\alias{tumor}
\docType{package}
\title{
Simulation of coalescence trees for tumour genomes.
}
\description{
Simulation of coalescence trees with coalescence times from a tumour model.
}
\details{
\tabular{ll}{
Package: \tab tumour\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2013-10-30\cr
License: \tab GPLv2\cr
}
%% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Thomas Mailund and Carsten Wiuf

Maintainer: Thomas Mailund <mailund@birc.au.dk>
}
\references{
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link[ape:ape-package]{ape}}
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
